+++ mobile/android/chrome/geckoview/geckoview.js
@@ -517,7 +517,8 @@ function startup() {
     {
       name: "GeckoViewContentBlocking",
       onEnable: {
-        resource: "resource://gre/modules/GeckoViewContentBlocking.jsm",
+        frameScript:
+          "chrome://geckoview/content/GeckoViewContentBlockingChild.js",
       },
     },
     {