+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/ContentBlocking.java
@@ -16,6 +16,7 @@ import android.support.annotation.IntDef;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.annotation.UiThread;
+import android.text.TextUtils;
 
 import org.mozilla.gecko.util.GeckoBundle;
 
@@ -36,15 +37,28 @@ public class ContentBlocking {
             }
 
             /**
-             * Set content blocking categories.
+             * Set anti-tracking categories.
              *
              * @param cat The categories of resources that should be blocked.
              *            Use one or more of the
-             *            {@link ContentBlocking#AT_AD AT_* or SB_*} flags.
+             *            {@link ContentBlocking.AntiTracking} flags.
              * @return This Builder instance.
              */
-            public @NonNull Builder categories(final @Category int cat) {
-                getSettings().setCategories(cat);
+            public @NonNull Builder antiTracking(final @CBAntiTracking int cat) {
+                getSettings().setAntiTracking(cat);
+                return this;
+            }
+
+            /**
+             * Set safe browsing categories.
+             *
+             * @param cat The categories of resources that should be blocked.
+             *            Use one or more of the
+             *            {@link ContentBlocking.SafeBrowsing} flags.
+             * @return This Builder instance.
+             */
+            public @NonNull Builder safeBrowsing(final @CBSafeBrowsing int cat) {
+                getSettings().setSafeBrowsing(cat);
                 return this;
             }
 
@@ -52,10 +66,10 @@ public class ContentBlocking {
              * Set cookie storage behavior.
              *
              * @param behavior The storage behavior that should be applied.
-             *                 Use one of the {@link #COOKIE_ACCEPT_ALL COOKIE_ACCEPT_*} flags.
+             *                 Use one of the {@link CookieBehavior} flags.
              * @return The Builder instance.
              */
-            public @NonNull Builder cookieBehavior(final @CookieBehavior int behavior) {
+            public @NonNull Builder cookieBehavior(final @CBCookieBehavior int behavior) {
                 getSettings().setCookieBehavior(behavior);
                 return this;
             }
@@ -64,10 +78,10 @@ public class ContentBlocking {
              * Set the cookie lifetime.
              *
              * @param lifetime The enforced cookie lifetime.
-             *                 Use one of the {@link #COOKIE_LIFETIME_NORMAL COOKIE_LIFETIME_*} flags.
+             *                 Use one of the {@link CookieLifetime} flags.
              * @return The Builder instance.
              */
-            public @NonNull Builder cookieLifetime(final @CookieLifetime int lifetime) {
+            public @NonNull Builder cookieLifetime(final @CBCookieLifetime int lifetime) {
                 getSettings().setCookieLifetime(lifetime);
                 return this;
             }
@@ -75,26 +89,26 @@ public class ContentBlocking {
 
         /* package */ final Pref<String> mAt = new Pref<String>(
             "urlclassifier.trackingTable",
-            ContentBlocking.catToAtPref(AT_DEFAULT));
+            ContentBlocking.catToAtPref(AntiTracking.DEFAULT));
         /* package */ final Pref<Boolean> mCm = new Pref<Boolean>(
             "privacy.trackingprotection.cryptomining.enabled", false);
         /* package */ final Pref<String> mCmList = new Pref<String>(
             "urlclassifier.features.cryptomining.blacklistTables",
-            ContentBlocking.catToCmListPref(NONE));
+            ContentBlocking.catToCmListPref(AntiTracking.NONE));
         /* package */ final Pref<Boolean> mFp = new Pref<Boolean>(
             "privacy.trackingprotection.fingerprinting.enabled", false);
         /* package */ final Pref<String> mFpList = new Pref<String>(
             "urlclassifier.features.fingerprinting.blacklistTables",
-            ContentBlocking.catToFpListPref(NONE));
+            ContentBlocking.catToFpListPref(AntiTracking.NONE));
 
         /* package */ final Pref<Boolean> mSbMalware = new Pref<Boolean>(
             "browser.safebrowsing.malware.enabled", true);
         /* package */ final Pref<Boolean> mSbPhishing = new Pref<Boolean>(
             "browser.safebrowsing.phishing.enabled", true);
         /* package */ final Pref<Integer> mCookieBehavior = new Pref<Integer>(
-            "network.cookie.cookieBehavior", COOKIE_ACCEPT_NON_TRACKERS);
+            "network.cookie.cookieBehavior", CookieBehavior.ACCEPT_NON_TRACKERS);
         /* package */ final Pref<Integer> mCookieLifetime = new Pref<Integer>(
-            "network.cookie.lifetimePolicy", COOKIE_LIFETIME_NORMAL);
+            "network.cookie.lifetimePolicy", CookieLifetime.NORMAL);
 
         /**
          * Construct default settings.
@@ -133,14 +147,14 @@ public class ContentBlocking {
         }
 
         /**
-         * Set content blocking categories.
+         * Set anti-tracking categories.
          *
          * @param cat The categories of resources that should be blocked.
          *            Use one or more of the
-         *            {@link ContentBlocking#AT_AD AT_* or SB_*} flags.
+         *            {@link ContentBlocking.AntiTracking} flags.
          * @return This Settings instance.
          */
-        public @NonNull Settings setCategories(final @Category int cat) {
+        public @NonNull Settings setAntiTracking(final @CBAntiTracking int cat) {
             mAt.commit(ContentBlocking.catToAtPref(cat));
 
             mCm.commit(ContentBlocking.catToCmPref(cat));
@@ -148,31 +162,50 @@ public class ContentBlocking {
 
             mFp.commit(ContentBlocking.catToFpPref(cat));
             mFpList.commit(ContentBlocking.catToFpListPref(cat));
+            return this;
+        }
 
+        /**
+         * Set safe browsing categories.
+         *
+         * @param cat The categories of resources that should be blocked.
+         *            Use one or more of the
+         *            {@link ContentBlocking.SafeBrowsing} flags.
+         * @return This Settings instance.
+         */
+        public @NonNull Settings setSafeBrowsing(final @CBSafeBrowsing int cat) {
             mSbMalware.commit(ContentBlocking.catToSbMalware(cat));
             mSbPhishing.commit(ContentBlocking.catToSbPhishing(cat));
             return this;
         }
 
         /**
-         * Get the set content blocking categories.
+         * Get the set anti-tracking categories.
          *
          * @return The categories of resources to be blocked.
          */
-        public @Category int getCategories() {
-            return ContentBlocking.atListToCat(mAt.get()) |
-                   ContentBlocking.cmListToCat(mCmList.get()) |
-                   ContentBlocking.fpListToCat(mFpList.get()) |
-                   ContentBlocking.sbMalwareToCat(mSbMalware.get()) |
-                   ContentBlocking.sbPhishingToCat(mSbPhishing.get());
+        public @CBAntiTracking int getAntiTrackingCategories() {
+            return ContentBlocking.atListToAtCat(mAt.get()) |
+                   ContentBlocking.cmListToAtCat(mCmList.get()) |
+                   ContentBlocking.fpListToAtCat(mFpList.get());
+        }
+
+        /**
+         * Get the set safe browsing categories.
+         *
+         * @return The categories of resources to be blocked.
+         */
+        public @CBAntiTracking int getSafeBrowsingCategories() {
+            return ContentBlocking.sbMalwareToSbCat(mSbMalware.get()) |
+                   ContentBlocking.sbPhishingToSbCat(mSbPhishing.get());
         }
 
         /**
          * Get the assigned cookie storage behavior.
          *
-         * @return The assigned behavior, as one of {@link #COOKIE_ACCEPT_ALL COOKIE_ACCEPT_*} flags.
+         * @return The assigned behavior, as one of {@link CookieBehavior} flags.
          */
-        public @CookieBehavior int getCookieBehavior() {
+        public @CBCookieBehavior int getCookieBehavior() {
             return mCookieBehavior.get();
         }
 
@@ -180,11 +213,11 @@ public class ContentBlocking {
          * Set cookie storage behavior.
          *
          * @param behavior The storage behavior that should be applied.
-         *                 Use one of the {@link #COOKIE_ACCEPT_ALL COOKIE_ACCEPT_*} flags.
+         *                 Use one of the {@link CookieBehavior} flags.
          * @return This Settings instance.
          */
         public @NonNull Settings setCookieBehavior(
-                final @CookieBehavior int behavior) {
+                final @CBCookieBehavior int behavior) {
             mCookieBehavior.commit(behavior);
             return this;
         }
@@ -192,9 +225,9 @@ public class ContentBlocking {
         /**
          * Get the assigned cookie lifetime.
          *
-         * @return The assigned lifetime, as one of {@link #COOKIE_LIFETIME_NORMAL COOKIE_LIFETIME_*} flags.
+         * @return The assigned lifetime, as one of {@link CookieLifetime} flags.
          */
-        public @CookieBehavior int getCookieLifetime() {
+        public @CBCookieLifetime int getCookieLifetime() {
             return mCookieLifetime.get();
         }
 
@@ -202,11 +235,11 @@ public class ContentBlocking {
          * Set the cookie lifetime.
          *
          * @param lifetime The enforced cookie lifetime.
-         *                 Use one of the {@link #COOKIE_LIFETIME_NORMAL COOKIE_LIFETIME_*} flags.
+         *                 Use one of the {@link CookieLifetime} flags.
          * @return This Settings instance.
          */
         public @NonNull Settings setCookieLifetime(
-                final @CookieLifetime int lifetime) {
+                final @CBCookieLifetime int lifetime) {
             mCookieLifetime.commit(lifetime);
             return this;
         }
@@ -227,166 +260,173 @@ public class ContentBlocking {
                 };
     }
 
+    public static class AntiTracking {
+        public static final int NONE = 0;
+
+        /**
+         * Block advertisement trackers.
+         */
+        public static final int AD = 1 << 1;
+
+        /**
+         * Block analytics trackers.
+         */
+        public static final int ANALYTIC = 1 << 2;
+
+        /**
+         * Block social trackers.
+         */
+        public static final int SOCIAL = 1 << 3;
+
+        /**
+         * Block content trackers.
+         * May cause issues with some web sites.
+         */
+        public static final int CONTENT = 1 << 4;
+
+        /**
+         * Block Gecko test trackers (used for tests).
+         */
+        public static final int TEST = 1 << 5;
+
+        /**
+         * Block cryptocurrency miners.
+         */
+        public static final int CRYPTOMINING = 1 << 6;
+
+        /**
+         * Block fingerprinting trackers.
+         */
+        public static final int FINGERPRINTING = 1 << 7;
+
+        /**
+         * Block ad, analytic, social and test trackers.
+         */
+        public static final int DEFAULT = AD | ANALYTIC | SOCIAL | TEST;
+
+        /**
+         * Block all known trackers.
+         * May cause issues with some web sites.
+         */
+        public static final int STRICT = DEFAULT | CONTENT | CRYPTOMINING | FINGERPRINTING;
+
+        protected AntiTracking() {}
+    }
+
     @Retention(RetentionPolicy.SOURCE)
     @IntDef(flag = true,
-            value = { NONE, AT_AD, AT_ANALYTIC, AT_SOCIAL, AT_CONTENT,
-                      AT_TEST, AT_CRYPTOMINING, AT_FINGERPRINTING,
-                      AT_DEFAULT, AT_STRICT,
-                      SB_MALWARE, SB_UNWANTED,
-                      SB_HARMFUL, SB_PHISHING,
-                      CB_DEFAULT, CB_STRICT })
-    /* package */ @interface Category {}
+            value = { AntiTracking.AD, AntiTracking.ANALYTIC,
+                      AntiTracking.SOCIAL, AntiTracking.CONTENT,
+                      AntiTracking.TEST, AntiTracking.CRYPTOMINING,
+                      AntiTracking.FINGERPRINTING,
+                      AntiTracking.DEFAULT, AntiTracking.STRICT,
+                      AntiTracking.NONE })
+    /* package */ @interface CBAntiTracking {}
 
-    public static final int NONE = 0;
+    public static class SafeBrowsing {
+        public static final int NONE = 0;
 
-    // Anti-tracking
-    /**
-     * Block advertisement trackers.
-     */
-    public static final int AT_AD = 1 << 1;
+        /**
+         * Block malware sites.
+         */
+        public static final int MALWARE = 1 << 10;
 
-    /**
-     * Block analytics trackers.
-     */
-    public static final int AT_ANALYTIC = 1 << 2;
+        /**
+         * Block unwanted sites.
+         */
+        public static final int UNWANTED = 1 << 11;
 
-    /**
-     * Block social trackers.
-     */
-    public static final int AT_SOCIAL = 1 << 3;
+        /**
+         * Block harmful sites.
+         */
+        public static final int HARMFUL = 1 << 12;
 
-    /**
-     * Block content trackers.
-     * May cause issues with some web sites.
-     */
-    public static final int AT_CONTENT = 1 << 4;
+        /**
+         * Block phishing sites.
+         */
+        public static final int PHISHING = 1 << 13;
 
-    /**
-     * Block Gecko test trackers (used for tests).
-     */
-    public static final int AT_TEST = 1 << 5;
+        /**
+         * Block all unsafe sites.
+         */
+        public static final int DEFAULT = MALWARE | UNWANTED | HARMFUL | PHISHING;
 
-    /**
-     * Block cryptocurrency miners.
-     */
-    public static final int AT_CRYPTOMINING = 1 << 6;
+        protected SafeBrowsing() {}
+    }
 
-    /**
-     * Block fingerprinting trackers.
-     */
-    public static final int AT_FINGERPRINTING = 1 << 7;
+    @Retention(RetentionPolicy.SOURCE)
+    @IntDef(flag = true,
+            value = { SafeBrowsing.MALWARE, SafeBrowsing.UNWANTED,
+                      SafeBrowsing.HARMFUL, SafeBrowsing.PHISHING,
+                      SafeBrowsing.DEFAULT, SafeBrowsing.NONE })
+    /* package */ @interface CBSafeBrowsing {}
 
-    /**
-     * Block ad, analytic, social and test trackers.
-     */
-    public static final int AT_DEFAULT =
-        AT_AD | AT_ANALYTIC | AT_SOCIAL | AT_TEST;
-
-    /**
-     * Block all known trackers.
-     * May cause issues with some web sites.
-     */
-    public static final int AT_STRICT =
-        AT_DEFAULT | AT_CONTENT | AT_CRYPTOMINING | AT_FINGERPRINTING;
-
-    // Safe browsing
-    /**
-     * Block malware sites.
-     */
-    public static final int SB_MALWARE = 1 << 10;
-
-    /**
-     * Block unwanted sites.
-     */
-    public static final int SB_UNWANTED = 1 << 11;
-
-    /**
-     * Block harmful sites.
-     */
-    public static final int SB_HARMFUL = 1 << 12;
-
-    /**
-     * Block phishing sites.
-     */
-    public static final int SB_PHISHING = 1 << 13;
-
-    /**
-     * Block all unsafe sites.
-     * Blocks all {@link #SB_MALWARE SB_*} types.
-     */
-    public static final int SB_ALL =
-        SB_MALWARE | SB_UNWANTED | SB_HARMFUL | SB_PHISHING;
-
-    /**
-     * Recommended content blocking categories.
-     * A combination of {@link #AT_DEFAULT} and {@link #SB_ALL}.
-     */
-    public static final int CB_DEFAULT = SB_ALL | AT_DEFAULT;
-
-    /**
-     * Strict content blocking categories.
-     * A combination of {@link #AT_STRICT} and {@link #SB_ALL}.
-     * May cause issues with some web sites.
-     */
-    public static final int CB_STRICT = SB_ALL | AT_STRICT;
 
     // Sync values with nsICookieService.idl.
+    public static class CookieBehavior {
+        /**
+         * Accept first-party and third-party cookies and site data.
+         */
+        public static final int ACCEPT_ALL = 0;
+
+        /**
+         * Accept only first-party cookies and site data to block cookies which are
+         * not associated with the domain of the visited site.
+         */
+        public static final int ACCEPT_FIRST_PARTY = 1;
+
+        /**
+         * Do not store any cookies and site data.
+         */
+        public static final int ACCEPT_NONE = 2;
+
+        /**
+         * Accept first-party and third-party cookies and site data only from
+         * sites previously visited in a first-party context.
+         */
+        public static final int ACCEPT_VISITED = 3;
+
+        /**
+         * Accept only first-party and non-tracking third-party cookies and site data
+         * to block cookies which are not associated with the domain of the visited
+         * site set by known trackers.
+         */
+        public static final int ACCEPT_NON_TRACKERS = 4;
+
+        protected CookieBehavior() {}
+    }
+
     @Retention(RetentionPolicy.SOURCE)
-    @IntDef({ COOKIE_ACCEPT_ALL, COOKIE_ACCEPT_FIRST_PARTY,
-              COOKIE_ACCEPT_NONE, COOKIE_ACCEPT_VISITED,
-              COOKIE_ACCEPT_NON_TRACKERS })
-    /* package */ @interface CookieBehavior {}
-
-    /**
-     * Accept first-party and third-party cookies and site data.
-     */
-    public static final int COOKIE_ACCEPT_ALL = 0;
-
-    /**
-     * Accept only first-party cookies and site data to block cookies which are
-     * not associated with the domain of the visited site.
-     */
-    public static final int COOKIE_ACCEPT_FIRST_PARTY = 1;
-
-    /**
-     * Do not store any cookies and site data.
-     */
-    public static final int COOKIE_ACCEPT_NONE = 2;
-
-    /**
-     * Accept first-party and third-party cookies and site data only from
-     * sites previously visited in a first-party context.
-     */
-    public static final int COOKIE_ACCEPT_VISITED = 3;
-
-    /**
-     * Accept only first-party and non-tracking third-party cookies and site data
-     * to block cookies which are not associated with the domain of the visited
-     * site set by known trackers.
-     */
-    public static final int COOKIE_ACCEPT_NON_TRACKERS = 4;
+    @IntDef({ CookieBehavior.ACCEPT_ALL, CookieBehavior.ACCEPT_FIRST_PARTY,
+              CookieBehavior.ACCEPT_NONE, CookieBehavior.ACCEPT_VISITED,
+              CookieBehavior.ACCEPT_NON_TRACKERS })
+    /* package */ @interface CBCookieBehavior {}
 
     // Sync values with nsICookieService.idl.
+    public static class CookieLifetime {
+        /**
+         * Accept default cookie lifetime.
+         */
+        public static final int NORMAL = 0;
+
+        /**
+         * Downgrade cookie lifetime to this runtime's lifetime.
+         */
+        public static final int RUNTIME = 2;
+
+        /**
+         * Limit cookie lifetime to N days.
+         * Defaults to 90 days.
+         */
+        public static final int DAYS = 3;
+
+        protected CookieLifetime() {}
+    }
+
     @Retention(RetentionPolicy.SOURCE)
-    @IntDef({ COOKIE_LIFETIME_NORMAL, COOKIE_LIFETIME_RUNTIME,
-              COOKIE_LIFETIME_DAYS })
-    /* package */ @interface CookieLifetime {}
-
-    /**
-     * Accept default cookie lifetime.
-     */
-    public static final int COOKIE_LIFETIME_NORMAL = 0;
-
-    /**
-     * Downgrade cookie lifetime to this runtime's lifetime.
-     */
-    public static final int COOKIE_LIFETIME_RUNTIME = 2;
-
-    /**
-     * Limit cookie lifetime to N days.
-     * Defaults to 90 days.
-     */
-    public static final int COOKIE_LIFETIME_DAYS = 3;
+    @IntDef({ CookieLifetime.NORMAL, CookieLifetime.RUNTIME,
+              CookieLifetime.DAYS })
+    /* package */ @interface CBCookieLifetime {}
 
     /**
      * Holds content block event details.
@@ -397,35 +437,82 @@ public class ContentBlocking {
          */
         public final @NonNull String uri;
 
-        /**
-         * The category types of the blocked resource.
-         * One or more of the {@link #AT_AD AT_* or SB_*} flags.
-         */
-        public final @Category int categories;
+        private final @CBAntiTracking int mAntiTrackingCat;
+        private final @CBSafeBrowsing int mSafeBrowsingCat;
+        private final @CBCookieBehavior int mCookieBehaviorCat;
+        private final boolean mIsBlocking;
 
         public BlockEvent(@NonNull final String uri,
-                          @Category final int categories) {
+                          final @CBAntiTracking int atCat,
+                          final @CBSafeBrowsing int sbCat,
+                          final @CBCookieBehavior int cbCat,
+                          final boolean isBlocking) {
             this.uri = uri;
-            this.categories = categories;
+            this.mAntiTrackingCat = atCat;
+            this.mSafeBrowsingCat = sbCat;
+            this.mCookieBehaviorCat = cbCat;
+            this.mIsBlocking = isBlocking;
         }
+
+        /**
+         * The anti-tracking category types of the blocked resource.
+         * One or more of the {@link AntiTracking} flags.
+         */
+        @UiThread
+        public @CBAntiTracking int getAntiTrackingCategory() {
+            return mAntiTrackingCat;
+        }
+
+        /**
+         * The safe browsing category types of the blocked resource.
+         * One or more of the {@link AntiTracking} flags.
+         */
+        @UiThread
+        public @CBSafeBrowsing int getSafeBrowsingCategory() {
+            return mSafeBrowsingCat;
+        }
+
+        /**
+         * The cookie types of the blocked resource.
+         * One or more of the {@link CookieBehavior} flags.
+         */
+        @UiThread
+        public @CBCookieBehavior int getCookieBehaviorCategory() {
+            return mCookieBehaviorCat;
+        }
+
 
         /* package */ static BlockEvent fromBundle(
                 @NonNull final GeckoBundle bundle) {
             final String uri = bundle.getString("uri");
-            final String matchedList = bundle.getString("matchedList");
+            final String blockedList = bundle.getString("blockedList");
+            final String loadedList =
+                TextUtils.join(",", bundle.getStringArray("loadedLists"));
             final long error = bundle.getLong("error", 0L);
+            final long category = bundle.getLong("category", 0L);
 
-            @Category int cats = NONE;
+            final String matchedList =
+                blockedList != null ? blockedList : loadedList;
 
-            if (matchedList != null) {
-                cats = ContentBlocking.atListToCat(matchedList) |
-                       ContentBlocking.cmListToCat(matchedList) |
-                       ContentBlocking.fpListToCat(matchedList);
-            } else if (error != 0L) {
-                cats = ContentBlocking.errorToCat(error);
-            }
+            final boolean blocking =
+                loadedList.isEmpty() &&
+                (blockedList != null ||
+                 error != 0L ||
+                 ContentBlocking.isBlockingGeckoCbCat(category));
 
-            return new BlockEvent(uri, cats);
+            return new BlockEvent(
+                uri,
+                ContentBlocking.atListToAtCat(matchedList) |
+                    ContentBlocking.cmListToAtCat(matchedList) |
+                    ContentBlocking.fpListToAtCat(matchedList),
+                ContentBlocking.errorToSbCat(error),
+                ContentBlocking.geckoCatToCbCat(category),
+                blocking);
+        }
+
+        @UiThread
+        public boolean isBlocking() {
+            return mIsBlocking;
         }
     }
 
@@ -445,8 +532,17 @@ public class ContentBlocking {
         @UiThread
         default void onContentBlocked(@NonNull GeckoSession session,
                                       @NonNull BlockEvent event) {}
-    }
 
+        /**
+         * A content element that could be blocked has been loaded.
+         *
+        * @param session The GeckoSession that initiated the callback.
+        * @param event The {@link BlockEvent} details.
+        */
+        @UiThread
+        default void onContentLoaded(@NonNull GeckoSession session,
+                                     @NonNull BlockEvent event) {}
+    }
 
     private static final String TEST = "moztest-track-simple";
     private static final String AD = "ads-track-digest256";
@@ -458,40 +554,40 @@ public class ContentBlocking {
     private static final String FINGERPRINTING =
         "base-fingerprinting-track-digest256";
 
-    /* package */ static @Category int sbMalwareToCat(final boolean enabled) {
-        return enabled ? (SB_MALWARE | SB_UNWANTED | SB_HARMFUL)
-                       : NONE;
+    /* package */ static @CBSafeBrowsing int sbMalwareToSbCat(final boolean enabled) {
+        return enabled ? (SafeBrowsing.MALWARE | SafeBrowsing.UNWANTED | SafeBrowsing.HARMFUL)
+                       : SafeBrowsing.NONE;
     }
 
-    /* package */ static @Category int sbPhishingToCat(final boolean enabled) {
-        return enabled ? SB_PHISHING
-                       : NONE;
+    /* package */ static @CBSafeBrowsing int sbPhishingToSbCat(final boolean enabled) {
+        return enabled ? SafeBrowsing.PHISHING
+                       : SafeBrowsing.NONE;
     }
 
-    /* package */ static boolean catToSbMalware(@Category final int cat) {
-        return (cat & (SB_MALWARE | SB_UNWANTED | SB_HARMFUL)) != 0;
+    /* package */ static boolean catToSbMalware(@CBAntiTracking final int cat) {
+        return (cat & (SafeBrowsing.MALWARE | SafeBrowsing.UNWANTED | SafeBrowsing.HARMFUL)) != 0;
     }
 
-    /* package */ static boolean catToSbPhishing(@Category final int cat) {
-        return (cat & SB_PHISHING) != 0;
+    /* package */ static boolean catToSbPhishing(@CBAntiTracking final int cat) {
+        return (cat & SafeBrowsing.PHISHING) != 0;
     }
 
-    /* package */ static String catToAtPref(@Category final int cat) {
+    /* package */ static String catToAtPref(@CBAntiTracking final int cat) {
         StringBuilder builder = new StringBuilder();
 
-        if ((cat & AT_TEST) != 0) {
+        if ((cat & AntiTracking.TEST) != 0) {
             builder.append(TEST).append(',');
         }
-        if ((cat & AT_AD) != 0) {
+        if ((cat & AntiTracking.AD) != 0) {
             builder.append(AD).append(',');
         }
-        if ((cat & AT_ANALYTIC) != 0) {
+        if ((cat & AntiTracking.ANALYTIC) != 0) {
             builder.append(ANALYTIC).append(',');
         }
-        if ((cat & AT_SOCIAL) != 0) {
+        if ((cat & AntiTracking.SOCIAL) != 0) {
             builder.append(SOCIAL).append(',');
         }
-        if ((cat & AT_CONTENT) != 0) {
+        if ((cat & AntiTracking.CONTENT) != 0) {
             builder.append(CONTENT).append(',');
         }
         if (builder.length() == 0) {
@@ -501,82 +597,127 @@ public class ContentBlocking {
         return builder.substring(0, builder.length() - 1);
     }
 
-    /* package */ static boolean catToCmPref(@Category final int cat) {
-        return (cat & AT_CRYPTOMINING) != 0;
+    /* package */ static boolean catToCmPref(@CBAntiTracking final int cat) {
+        return (cat & AntiTracking.CRYPTOMINING) != 0;
     }
 
-    /* package */ static String catToCmListPref(@Category final int cat) {
+    /* package */ static String catToCmListPref(@CBAntiTracking final int cat) {
         StringBuilder builder = new StringBuilder();
 
-        if ((cat & AT_CRYPTOMINING) != 0) {
+        if ((cat & AntiTracking.CRYPTOMINING) != 0) {
             builder.append(CRYPTOMINING);
         }
         return builder.toString();
     }
 
-    /* package */ static boolean catToFpPref(@Category final int cat) {
-        return (cat & AT_FINGERPRINTING) != 0;
+    /* package */ static boolean catToFpPref(@CBAntiTracking final int cat) {
+        return (cat & AntiTracking.FINGERPRINTING) != 0;
     }
 
-    /* package */ static String catToFpListPref(@Category final int cat) {
+    /* package */ static String catToFpListPref(@CBAntiTracking final int cat) {
         StringBuilder builder = new StringBuilder();
 
-        if ((cat & AT_FINGERPRINTING) != 0) {
+        if ((cat & AntiTracking.FINGERPRINTING) != 0) {
             builder.append(FINGERPRINTING);
         }
         return builder.toString();
     }
 
-    /* package */ static @Category int fpListToCat(final String list) {
-        int cat = 0;
+    /* package */ static @CBAntiTracking int fpListToAtCat(final String list) {
+        int cat = AntiTracking.NONE;
+        if (list == null) {
+            return cat;
+        }
         if (list.indexOf(FINGERPRINTING) != -1) {
-            cat |= AT_FINGERPRINTING;
+            cat |= AntiTracking.FINGERPRINTING;
         }
         return cat;
     }
 
-    /* package */ static @Category int atListToCat(final String list) {
-        int cat = 0;
+    /* package */ static @CBAntiTracking int atListToAtCat(final String list) {
+        int cat = AntiTracking.NONE;
+
+        if (list == null) {
+            return cat;
+        }
         if (list.indexOf(TEST) != -1) {
-            cat |= AT_TEST;
+            cat |= AntiTracking.TEST;
         }
         if (list.indexOf(AD) != -1) {
-            cat |= AT_AD;
+            cat |= AntiTracking.AD;
         }
         if (list.indexOf(ANALYTIC) != -1) {
-            cat |= AT_ANALYTIC;
+            cat |= AntiTracking.ANALYTIC;
         }
         if (list.indexOf(SOCIAL) != -1) {
-            cat |= AT_SOCIAL;
+            cat |= AntiTracking.SOCIAL;
         }
         if (list.indexOf(CONTENT) != -1) {
-            cat |= AT_CONTENT;
+            cat |= AntiTracking.CONTENT;
         }
         return cat;
     }
 
-    /* package */ static @Category int cmListToCat(final String list) {
-        int cat = 0;
+    /* package */ static @CBAntiTracking int cmListToAtCat(final String list) {
+        int cat = AntiTracking.NONE;
+        if (list == null) {
+            return cat;
+        }
         if (list.indexOf(CRYPTOMINING) != -1) {
-            cat |= AT_CRYPTOMINING;
+            cat |= AntiTracking.CRYPTOMINING;
         }
         return cat;
     }
 
-    /* package */ static @Category int errorToCat(final long error) {
+    /* package */ static @CBSafeBrowsing int errorToSbCat(final long error) {
         // Match flags with XPCOM ErrorList.h.
         if (error == 0x805D001FL) {
-            return SB_PHISHING;
+            return SafeBrowsing.PHISHING;
         }
         if (error == 0x805D001EL) {
-            return SB_MALWARE;
+            return SafeBrowsing.MALWARE;
         }
         if (error == 0x805D0023L) {
-            return SB_UNWANTED;
+            return SafeBrowsing.UNWANTED;
         }
         if (error == 0x805D0026L) {
-            return SB_HARMFUL;
+            return SafeBrowsing.HARMFUL;
         }
-        return NONE;
+        return SafeBrowsing.NONE;
+    }
+
+    // Match flags with nsIWebProgressListener.idl.
+    private static final long STATE_COOKIES_LOADED = 0x8000L;
+    private static final long STATE_COOKIES_BLOCKED_TRACKER = 0x20000000L;
+    private static final long STATE_COOKIES_BLOCKED_ALL = 0x40000000L;
+    private static final long STATE_COOKIES_BLOCKED_FOREIGN = 0x80L;
+
+    /* package */ static boolean isBlockingGeckoCbCat(final long geckoCat) {
+        return
+            (geckoCat &
+                (STATE_COOKIES_BLOCKED_TRACKER |
+                 STATE_COOKIES_BLOCKED_ALL |
+                 STATE_COOKIES_BLOCKED_FOREIGN))
+            != 0;
+    }
+
+    /* package */ static @CBCookieBehavior int geckoCatToCbCat(
+            final long geckoCat) {
+        if ((geckoCat & STATE_COOKIES_LOADED) != 0) {
+            // We don't know which setting would actually block this cookie, so
+            // we return the most strict value.
+            return CookieBehavior.ACCEPT_NONE;
+        }
+        if ((geckoCat & STATE_COOKIES_BLOCKED_FOREIGN) != 0) {
+            return CookieBehavior.ACCEPT_FIRST_PARTY;
+        }
+        if ((geckoCat & STATE_COOKIES_BLOCKED_TRACKER) != 0) {
+            return CookieBehavior.ACCEPT_NON_TRACKERS;
+        }
+        if ((geckoCat & STATE_COOKIES_BLOCKED_ALL) != 0) {
+            return CookieBehavior.ACCEPT_NONE;
+        }
+        // TODO: There are more reasons why cookies may be blocked.
+        return CookieBehavior.ACCEPT_ALL;
     }
 }